put MapsGrid in explicit file (and some code restructuring)
